pyproject.toml
[project]
name = "guitar-safari"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "aiohttp[speedups] (>=3.12.15,<4.0.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "sqlalchemy[postgresql-psycopg] (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)"
]

[tool.poetry]
packages = [{include = "guitar_safari", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
